# docs/openapi.yaml
openapi: 3.1.0
info:
  title: edu-platform
  version: 0.1.0
paths:
...
  /v1/auth/otp/request:
    post:
      tags:
      - auth
      summary: Request Otp
      operationId: request_otp_v1_auth_otp_request_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpRequestIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpRequestOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/auth/otp/verify:
    post:
      tags:
      - auth
      summary: Verify Otp
      operationId: verify_otp_v1_auth_otp_verify_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpVerifyIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/auth/refresh:
    post:
      tags:
      - auth
      summary: Refresh
      operationId: refresh_v1_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/auth/logout:
    post:
      tags:
      - auth
      summary: Logout
      operationId: logout_v1_auth_logout_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/auth/me:
    get:
      tags:
      - auth
      summary: Me
      operationId: me_v1_auth_me_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - auth
      summary: Update Me
      operationId: update_me_v1_auth_me_patch
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdateIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/me:
    get:
      tags:
      - ranking
      summary: Me Overview
      operationId: me_overview_v1_me_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/me/coins:
    get:
      tags:
      - ranking
      summary: My Coins
      operationId: my_coins_v1_me_coins_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OtpRequestIn:
      properties:
        phone:
          type: string
          title: Phone
        role:
          type: string
          title: Role
          default: student
      type: object
      required:
      - phone
      title: OtpRequestIn
    OtpRequestOut:
      properties:
        retry_after_seconds:
          type: integer
          title: Retry After Seconds
        expires_in_seconds:
          type: integer
          title: Expires In Seconds
        debug_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Debug Code
      type: object
      required:
      - retry_after_seconds
      - expires_in_seconds
      title: OtpRequestOut
    OtpVerifyIn:
      properties:
        phone:
          type: string
          title: Phone
        code:
          type: string
          maxLength: 8
          minLength: 4
          title: Code
        display_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Display Name
        region_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Region Code
        grade:
          anyOf:
          - type: integer
          - type: 'null'
          title: Grade
      type: object
      required:
      - phone
      - code
      title: OtpVerifyIn
    ProfileUpdateIn:
      properties:
        display_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Display Name
        region_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Region Code
        grade:
          anyOf:
          - type: integer
          - type: 'null'
          title: Grade
        locale:
          anyOf:
          - type: string
          - type: 'null'
          title: Locale
      type: object
      title: ProfileUpdateIn
...
    RefreshIn:
      properties:
        refresh_token:
          type: string
          title: Refresh Token
      type: object
      required:
      - refresh_token
      title: RefreshIn
...
    TokenPair:
      properties:
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        expires_in:
          type: integer
          title: Expires In
      type: object
      required:
      - access_token
      - refresh_token
      - expires_in
      title: TokenPair
    UserOut:
      properties:
        id:
          type: string
          title: Id
        role:
          type: string
          title: Role
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        display_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Display Name
        region_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Region Code
        grade:
          anyOf:
          - type: integer
          - type: 'null'
          title: Grade
        locale:
          anyOf:
          - type: string
          - type: 'null'
          title: Locale
      type: object
      required:
      - id
      - role
      title: UserOut
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError